The service to register a new user from the view is created by connecting the frontend and backend.
