+++ backend/package.json
@@ -28,6 +28,7 @@
     "@nestjs/typeorm": "^11.0.0",
     "class-transformer": "0.5.1",
     "class-validator": "0.14.1",
+    "firebase-admin": "^13.3.0",
     "pg": "^8.15.5",
     "reflect-metadata": "^0.2.2",
     "rxjs": "^7.8.1",
