+++ frontend/src/App.tsx
@@ -1,9 +1,9 @@
 import { BrowserRouter as Router, Routes, Route } from 'react-router-dom';
 import Login from './assets/pages/Login';
 import Home from './assets/pages/Home';
-import Register from './assets/pages/Register';
 import './index.css';
 import './App.css';
+import RegisterCli from './assets/pages/RegisterCli';
 
 function App() {
   return (
@@ -12,7 +12,7 @@ function App() {
         <Routes>
           <Route path="/" element={<Home />} />
           <Route path="/login" element={<Login />} />
-          <Route path="/register" element={<Register />} />
+          <Route path="/register" element={<RegisterCli />} />
           {/* <Route path="/benefits" element={<Benefits />} /> */}
         </Routes>
       </div>
